refactor: organize APIs
